Put Perl run files in _common consistently
In the `ThirdMigration` tutorial part, I noticed that the Perl run files
weren't put into the `share/migrations/_common` path as had been the
case in earlier tutorial parts.  This seemed inconsistent with the rest
of the tutorial and surprised me as a reader, hence this change alters
the paths of these files to be in the `_common` path (and adds a command
to create the appropriate `upgrade` directory).  The change has also
been tested to ensure that it still works as intended within the
tutorial.